// frontend/src/types/exercise-set.types.tsx
export interface ExerciseSetType {
   exerciseSetId?: number;
   exerciseSetOrder: number;
   hasReps: boolean;
   isTimed: boolean;
   isDistance: boolean;
   reps: number;
   weight: number;
   weightUnit: number;
   hr: number;
   min: number;
   sec: number;
   distance: number;
   distanceUnit: number;
}